fix(data): supply correct class labels and remove redundant label encoder refactor(data): move config file to src/config dir
