+++ src/configs/prithvi_config.yaml
